requirements.txt
streamlit
torch
torchvision
facenet-pytorch
numpy
pillow
tensorflow
opencv-python-headless
librosa
face-recognition
